tfc/requirement.txt
Pillow==9.3.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

